requirements/base.txt
Django==3.0.7
djangorestframework==3.11.0
feedparser==5.2.1
